Convert h.api module into a package
This is a refactoring of the h.api module into a package.

- the get_user helper moves to h.api.auth
- elasticsearch setup moves to h.api.db
- views move to h.api.views
