+++ h/api/models.py
@@ -1,0 +1,9 @@
+# -*- coding: utf-8 -*-
+
+from ..models import Annotation
+from ..models import Document
+
+__all__ = [
+    'Annotation',
+    'Document'
+]
